Extract Card types
