+++ src/components/Card/Card.tsx
@@ -1,41 +1,15 @@
 import { h } from 'preact';
-import { UIElement, UIElementProps } from '../UIElement';
+import { UIElement } from '../UIElement';
 import CardInput from './components/CardInput';
 import CoreProvider from '~/core/Context/CoreProvider';
 import getImage from '~/utils/get-image';
 import collectBrowserInfo from '~/utils/browserInfo';
 import fetchJSONData from '~/utils/fetch-json-data';
-
-export interface CardElementProps extends UIElementProps {
-    type?: string;
-    brand?: string;
-
-    /** @deprecated use brands instead */
-    groupTypes?: string[];
-
-    brands?: string[];
-    enableStoreDetails?: boolean;
-    hideCVC?: boolean;
-    hasHolderName?: boolean;
-    holderNameRequired?: boolean;
-    [key: string]: any;
-}
-
-interface CardElementData {
-    paymentMethod: any;
-    billingAddress?: any;
-    installments?: any;
-    storePaymentMethod?: any;
-    browserInfo: any;
-}
+import { CardElementData, CardElementProps } from './types';
 
 export class CardElement extends UIElement<CardElementProps> {
     public static type = 'scheme';
     private currentRequestId;
-
-    constructor(props: CardElementProps) {
-        super(props);
-    }
 
     formatProps(props: CardElementProps) {
         return {
